Fix regression leading to duplicates
Due to the shift in columns, we were encountering duplicate events being
added again and again, in addition, the special characters in the events
name led to duplicates being added despite the new logic. This commit
fixes that, however, more care should be taken in how input from the
source is parsed.